feat: Modify navie-ui provide
