+++ src/App.vue
@@ -1,23 +1,11 @@
 <script setup lang="ts">
-import { NConfigProvider, zhTW, dateZhTW } from 'naive-ui';
 import NaiveUIProvider from './components/common/NaiveUIProvider.vue';
-import { useThemeStore } from '@/stores';
-
-const theme = useThemeStore();
 </script>
 
 <template>
-  <n-config-provider
-    class="h-full"
-    :locale="zhTW"
-    :data-locale="dateZhTW"
-    :preflight-style-disable="true"
-    :theme="theme.naiveTheme"
-  >
-    <NaiveUIProvider>
-      <RouterView />
-    </NaiveUIProvider>
-  </n-config-provider>
+  <NaiveUIProvider>
+    <RouterView />
+  </NaiveUIProvider>
 </template>
 
 <style scoped></style>
