+++ src/components/common/NaiveUIProvider.vue
@@ -1,18 +1,40 @@
 <template>
-  <n-loading-bar-provider>
-    <slot></slot>
-    <NaiveContentProvider />
-  </n-loading-bar-provider>
+  <n-config-provider
+    class="h-full"
+    :locale="zhTW"
+    :data-locale="dateZhTW"
+    :preflight-style-disable="true"
+    :theme="theme.naiveTheme"
+  >
+    <n-loading-bar-provider>
+      <n-notification-provider>
+        <slot></slot>
+        <NaiveContentProvider />
+      </n-notification-provider>
+    </n-loading-bar-provider>
+  </n-config-provider>
 </template>
 
 <script setup lang="ts">
 import { h, defineComponent } from 'vue';
-import { useLoadingBar, NLoadingBarProvider } from 'naive-ui';
+import {
+  NConfigProvider,
+  NLoadingBarProvider,
+  NNotificationProvider,
+  zhTW,
+  dateZhTW,
+  useLoadingBar,
+  useNotification
+} from 'naive-ui';
+import { useThemeStore } from '@/stores';
+
+const theme = useThemeStore();
 
 const NaiveContentProvider = defineComponent({
   name: 'NaiveContentProvider',
   setup() {
-    window.$loading = useLoadingBar();
+    window.$loadingBar = useLoadingBar();
+    window.$notify = useNotification();
 
     return () => h('div');
   }
