+++ src/types/global.d.ts
@@ -1,0 +1,4 @@
+declare interface Window {
+  $loadingBar: import('naive-ui').LoadingBarApi;
+  $notify: import('naive-ui').NotificationApi;
+}
